+++ lib/src/runtime/dio/dio_connection_pool.dart
@@ -1,0 +1,2 @@
+export 'dio_connection_pool_stub.dart'
+    if (dart.library.io) 'dio_connection_pool_io.dart';